+++ integrations/gitlab-v2/tests/clients/test_gitlab_client.py
@@ -176,9 +176,7 @@ class TestGitLabClient:
 
             # Assert
             assert result == mock_group
-            mock_send_request.assert_called_once_with(
-                "GET", f"groups/{group_id}", params=client.DEFAULT_PARAMS
-            )
+            mock_send_request.assert_called_once_with("GET", f"groups/{group_id}")
 
     async def test_get_merge_request(self, client: GitLabClient) -> None:
         """Test fetching a single merge request by ID"""
@@ -331,11 +329,9 @@ class TestGitLabClient:
             {"id": 2, "name": "Project 2"},
         ]
 
-        async def mock_enrich_func(
-            project: dict[str, Any]
-        ) -> AsyncGenerator[list[dict[str, Any]], None]:
+        async def mock_enrich_func(project: dict[str, Any]) -> dict[str, Any]:
             project["enriched"] = True
-            yield [project]
+            return project
 
         # Act
         result = await client._enrich_batch(batch, mock_enrich_func, max_concurrent=1)
@@ -344,3 +340,202 @@ class TestGitLabClient:
         assert len(result) == 2
         assert result[0]["enriched"] is True
         assert result[1]["enriched"] is True
+
+    async def test_search_files_in_repos(self, client: GitLabClient) -> None:
+        """Test file search in specific repositories using scope and query via _search_in_repository"""
+        processed_files = [
+            {"path": "test.json", "project_id": "123", "content": {"key": "value"}}
+        ]
+        repos = ["group/project"]
+        scope = "blobs"
+        query = "test.json"
+        with patch.object(
+            client,
+            "_search_files_in_repository",
+            return_value=async_mock_generator([processed_files]),
+        ) as mock_search_repo:
+            with patch.object(
+                client, "get_file_content", return_value='{"key": "value"}'
+            ):
+                results = []
+                async for batch in client.search_files(
+                    scope, query, repositories=repos, skip_parsing=False
+                ):
+                    results.extend(batch)
+                assert len(results) == 1
+                assert results[0]["path"] == "test.json"
+                assert results[0]["content"] == {"key": "value"}
+                mock_search_repo.assert_called_once_with(
+                    "group/project", "blobs", "path:test.json", False
+                )
+
+    async def test_search_files_in_groups(self, client: GitLabClient) -> None:
+        """Test file search across all groups using scope and query"""
+        mock_groups = [{"id": "1", "name": "Group1"}]
+        processed_files = [
+            {"path": "test.yaml", "project_id": "123", "content": {"key": "value"}}
+        ]
+        scope = "blobs"
+        query = "test.yaml"
+        with patch.object(
+            client, "get_groups", return_value=async_mock_generator([mock_groups])
+        ):
+            with patch.object(
+                client,
+                "_search_files_in_group",
+                return_value=async_mock_generator([processed_files]),
+            ) as mock_search_group:
+                with patch.object(
+                    client, "get_file_content", return_value="key: value"
+                ):
+                    results = []
+                    async for batch in client.search_files(
+                        scope, query, skip_parsing=False
+                    ):
+                        results.extend(batch)
+                    assert len(results) == 1
+                    assert results[0]["path"] == "test.yaml"
+                    assert results[0]["content"] == {"key": "value"}
+                    mock_search_group.assert_called_once_with(
+                        "1", "blobs", "path:test.yaml", False
+                    )
+
+    async def test_get_file_content(self, client: GitLabClient) -> None:
+        """Test fetching file content via REST"""
+        # Arrange
+        project_id = "123"
+        file_path = "test.txt"
+        mock_content = "Hello, World!"
+        with patch.object(
+            client.rest,
+            "get_file_content",
+            return_value=mock_content,
+        ) as mock_get_file_content:
+            # Act
+            result = await client.get_file_content(project_id, file_path, "main")
+
+            # Assert
+            assert result == mock_content
+            mock_get_file_content.assert_called_once_with(project_id, file_path, "main")
+
+    async def test_get_project(self, client: GitLabClient) -> None:
+        """Test fetching a single project by path"""
+        project_path = "group/project"
+        mock_project = {
+            "id": "123",
+            "path_with_namespace": project_path,
+            "default_branch": "main",
+        }
+        with patch.object(
+            client.rest,
+            "send_api_request",
+            AsyncMock(return_value=mock_project),
+        ) as mock_send_request:
+            result = await client.get_project(project_path)
+            assert result["id"] == "123"
+            assert result["path_with_namespace"] == project_path
+            mock_send_request.assert_called_once_with("GET", "projects/group%2Fproject")
+
+    async def test_file_exists(self, client: GitLabClient) -> None:
+        """Test checking if a file exists in a project"""
+        project_id = "123"
+        scope = "blobs"
+        query = "test.txt"
+        mock_response = [{"path": "test.txt"}]  # Non-empty response means exists
+        with patch.object(
+            client.rest,
+            "send_api_request",
+            AsyncMock(return_value=mock_response),
+        ) as mock_send_request:
+            exists = await client.file_exists(project_id, scope, query)
+            assert exists is True
+            mock_send_request.assert_called_once_with(
+                "GET",
+                "projects/123/search",
+                params={"scope": "blobs", "search": "test.txt"},
+            )
+
+        # Test non-existent file
+        with patch.object(
+            client.rest,
+            "send_api_request",
+            AsyncMock(return_value=[]),  # Empty response means doesn't exist
+        ) as mock_send_request:
+            exists = await client.file_exists(project_id, scope, query)
+            assert exists is False
+
+    async def test_get_repository_tree(self, client: GitLabClient) -> None:
+        """Test fetching repository tree (folders only) for a project"""
+        project = {"path_with_namespace": "group/project"}
+        path = "src"
+        ref = "main"
+        mock_tree = [
+            {"type": "tree", "name": "folder1"},
+            {"type": "blob", "name": "file.txt"},
+            {"type": "tree", "name": "folder2"},
+        ]
+        with patch.object(
+            client.rest,
+            "get_paginated_project_resource",
+            return_value=async_mock_generator([mock_tree]),
+        ) as mock_get_paginated:
+            results = []
+            async for batch in client.get_repository_tree(project, path, ref):
+                results.extend(batch)
+
+            assert len(results) == 2
+            assert results[0]["folder"]["name"] == "folder1"
+            assert results[1]["folder"]["name"] == "folder2"
+            assert all(r["repo"] == project for r in results)
+            assert all(r["__branch"] == ref for r in results)
+            mock_get_paginated.assert_called_once_with(
+                "group/project",
+                "repository/tree",
+                {"ref": "main", "path": "src", "recursive": False},
+            )
+
+    async def test_get_repository_folders(self, client: GitLabClient) -> None:
+        """Test searching folders in a single repository"""
+        # Arrange
+        repository = "group/project1"
+        path = "src"
+        branch = "develop"
+
+        mock_project = {
+            "id": "1",
+            "path_with_namespace": "group/project1",
+            "default_branch": "main",
+        }
+
+        mock_tree = [
+            {"type": "tree", "name": "folder1"},
+            {"type": "blob", "name": "file.txt"},
+        ]
+
+        with patch.object(
+            client, "get_project", AsyncMock(return_value=mock_project)
+        ) as mock_get_project:
+            with patch.object(
+                client.rest,
+                "get_paginated_project_resource",
+                return_value=async_mock_generator([mock_tree]),
+            ) as mock_get_paginated:
+                # Act
+                results = []
+                async for batch in client.get_repository_folders(
+                    path, repository, branch
+                ):
+                    results.extend(batch)
+
+                # Assert
+                assert len(results) == 1  # Only one folder from mock_tree
+                assert results[0]["folder"]["name"] == "folder1"
+                assert results[0]["repo"] == mock_project
+                assert results[0]["__branch"] == "develop"
+
+                mock_get_project.assert_called_once_with("group/project1")
+                mock_get_paginated.assert_called_once_with(
+                    "group/project1",
+                    "repository/tree",
+                    {"ref": "develop", "path": "src", "recursive": False},
+                )
